ci: 🎡 determine relevant podspec version from changed files
Previous way to determine the relevant SDK version was incorrect. When
introducing 6.x podfiles then the determined version was 7.x. With this
commit the version will be determined by parsing the path(s) of the
changed file(s) of the commit that triggered the GitHub workflow run. In
case no podspec file was chagned/added then a fallback version (7.05) is
used.